app/requirements.txt
Flask==1.1.2
werkzeug==0.16.0
requests

# Application server for both development and production.
gunicorn==20.0.*

# Testing and static analysis.
pytest==5.3.2
pytest-cov==2.8.*
flake8==3.7.*
pytz==2018.5

# Forms.
Flask-WTF==0.14.*
WTForms-Components==0.10.*

# Extensions.
Flask-Admin==1.5.6
flask-debugtoolbar==0.10.0
flask-less==0.4
flask-oidc==1.4.0
Flask-Limiter==1.1.*
python-memcached==1.59
redis==3.2.1

# Supervisor.
psutil==5.4.8
watchdog==0.9.0
